Introduce new configurations for runtime
This commits adds 3 new configurations aimed at bringing a better "runtime" support for both
the `java` and `java-library` plugins:

- `runtimeOnly` is a configuration where the user would add dependencies that are only required at runtime.
There are multiple uses for that, but it typically includes services or JDBC drivers. The runtime is responsible
for loading them, and they are not required at runtime. In addition, the dependencies of this bucket are *not*
exported to consumers. So the runtime of a consumer will not be polluted by the `runtimeOnly` dependencies of
a dependency.
- `runtimeElements` is a configuration that is _consumable_. It contains the dependencies that are strictly
required by the component at runtime. They are exported to consumers, because without them the program would simply
not run.
- `runtimeClasspath` is a configuration that extends from `runtimeOnly` and `runtimeElements` and contains all
runtime elements that are required for this component to be used independently of any consumer.

This typically solves the problem of programs depending on `slf4j-api` at compile time, but need an implementation
at some point, so pollute consumers with this implementation (which has to be excluded).