Reset performance test baselines
Performance tests are automatically run against the last release,
there is no need to specify that explicitly. Only if there was an
acceptable regression should we explicitly specify a nightly.

We'll probably have to add a few of those nightly versions back in,
but many of them were just added in an attempt to stabilize tests,
not because of actual regressions.

Regressions that we accepted but might want to get back to will be
handled by the historical performance tests.